hide osquery/events/linux/ebpf/system.h from the external header: other code will not need it
