+++ osquery/events/CMakeLists.txt
@@ -44,9 +44,10 @@ else()
     "${CMAKE_CURRENT_LIST_DIR}/linux/auditdnetlink.h"
     "${CMAKE_CURRENT_LIST_DIR}/linux/auditeventpublisher.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/linux/auditeventpublisher.h"
+    "${CMAKE_CURRENT_LIST_DIR}/linux/ebpf/ebpf.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/linux/ebpf/ebpf.h"
-    "${CMAKE_CURRENT_LIST_DIR}/linux/ebpf/system.h"
     "${CMAKE_CURRENT_LIST_DIR}/linux/ebpf/system.cpp"
+    "${CMAKE_CURRENT_LIST_DIR}/linux/ebpf/system.h"
     "${CMAKE_CURRENT_LIST_DIR}/linux/inotify.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/linux/inotify.h"
     "${CMAKE_CURRENT_LIST_DIR}/linux/syslog.cpp"
